requirements.txt
amqp==2.5.1
beautifulsoup4==4.8.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
coverage==4.5.4
dj-database-url==0.5.0
Django==2.2.5
django-heroku==0.3.1
django-robots==3.1.0
django-seo==0.3.5
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
kombu==4.6.4
more-itertools==7.2.0
Pillow==6.1.0
psycopg2==2.8.3
python-decouple==3.1
python-memcached==1.59
pytz==2019.2
requests==2.22.0
sentry-sdk==0.12.2
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
urllib3==1.25.6
vine==1.3.0
whitenoise==4.1.4
zipp==0.6.0
